Much faster surface fitting (from Will)
Using matlab linear equation solving instead of the heavy fit function, making it now feasible to fit the entire 37000*32000 dataset! However, if we also do partial differentiations to find aspect, it is 10 times slower. So probably that job and slope analysis would be better for gdaldem. Through slope and aspect data we may be able to come back to MATLAB and run convexity analysis, etc.